boards: cogip2019-cortex-native: move _set_shm_key
As this command is specific to native architecture for emulation,
move it to the cogip2019-cortex-native board.
